# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PicDefense.io APIv2",
    "description": "API Documentation for PicDefense.io \n\n All endpoints require an API TOKEN in header 'X-API-TOKEN'. \n\n Your X-API-TOKEN is user id and api key seperated by a colon. Example: USERID:APIKEY \n\n Please visit the <a href=\"/dashboard/settings\" target=\"_blank\">Settings</a> page for these details.",
    "termsOfService": "https://picdefense.io/terms-conditions/",
    "contact": {
      "name": "API Support",
      "url": "https://help.picdefense.io/"
    },
    "version": "2"
  },
...
  "servers": [
    {
      "url": "https://app.picdefense.io/api/v2"
    }
  ],
...
  "paths": {
    "/credits": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "remaining account credits",
        "operationId": "credits",
        "responses": {
          "200": {
            "description": "succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": 1
                    },
                    "message": {
                      "type": "string",
                      "example": "succeeded"
                    },
                    "error": {
                      "type": "string",
                      "example": "none"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "credits": {
                          "type": "integer",
                          "format": "int32",
                          "example": 100
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "failed - unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unauthorized"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "failed - unknown server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unknown server error"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
    "/landmark": {
      "post": {
        "tags": [
          "landmark"
        ],
        "summary": "check if image contains a landmark",
        "description": "Returns whether the image contains a landmark and its name if found. Consumes 0.03 credits on success.",
        "operationId": "landmark",
        "requestBody": {
          "description": "image URL",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "url",
                    "example": "http://example.com/image.jpg",
                    "description": "The URL of the image to check. Must be publicly accessible."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": 1
                    },
                    "message": {
                      "type": "string",
                      "example": "succeeded"
                    },
                    "error": {
                      "type": "string",
                      "example": "none"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "landmark": {
                          "type": "boolean",
                          "example": true,
                          "description": "Whether the image contains a landmark."
                        },
                        "description": {
                          "type": "string",
                          "example": "Colosseum",
                          "description": "The name of the detected landmark, if any."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "failed - unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unauthorized"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "failed - insufficient credits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "insufficient credits"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "failed - unknown server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unknown server error"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
    "/safesearch": {
      "post": {
        "tags": [
          "safesearch"
        ],
        "summary": "check image content safety",
        "description": "Returns content safety analysis of the image. Consumes 0.03 credits on success.",
        "operationId": "safesearch",
        "requestBody": {
          "description": "image URL",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "url",
                    "example": "http://example.com/image.jpg",
                    "description": "The URL of the image to check. Must be publicly accessible."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": 1
                    },
                    "message": {
                      "type": "string",
                      "example": "succeeded"
                    },
                    "error": {
                      "type": "string",
                      "example": "none"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "safesearch": {
                          "type": "object",
                          "properties": {
                            "adult": {
                              "type": "string",
                              "example": "VERY_UNLIKELY",
                              "description": "Likelihood of adult content",
                              "enum": ["VERY_UNLIKELY", "UNLIKELY", "POSSIBLE", "LIKELY", "VERY_LIKELY"]
                            },
                            "spoof": {
                              "type": "string",
                              "example": "VERY_UNLIKELY",
                              "description": "Likelihood of spoofed content",
                              "enum": ["VERY_UNLIKELY", "UNLIKELY", "POSSIBLE", "LIKELY", "VERY_LIKELY"]
                            },
                            "medical": {
                              "type": "string",
                              "example": "VERY_UNLIKELY",
                              "description": "Likelihood of medical content",
                              "enum": ["VERY_UNLIKELY", "UNLIKELY", "POSSIBLE", "LIKELY", "VERY_LIKELY"]
                            },
                            "violence": {
                              "type": "string",
                              "example": "VERY_UNLIKELY",
                              "description": "Likelihood of violent content",
                              "enum": ["VERY_UNLIKELY", "UNLIKELY", "POSSIBLE", "LIKELY", "VERY_LIKELY"]
                            },
                            "racy": {
                              "type": "string",
                              "example": "VERY_UNLIKELY",
                              "description": "Likelihood of racy content",
                              "enum": ["VERY_UNLIKELY", "UNLIKELY", "POSSIBLE", "LIKELY", "VERY_LIKELY"]
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "failed - unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unauthorized"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "failed - insufficient credits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "insufficient credits"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "failed - unknown server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unknown server error"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
    "/watermark": {
      "post": {
        "tags": [
          "watermark"
        ],
        "summary": "detect a visible watermark",
        "description": "Detects a visible stock/photographer watermark in the image and identifies its source. Consumes 0.03 credits on success.",
        "operationId": "watermark",
        "requestBody": {
          "description": "image URL",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "url",
                    "example": "http://example.com/image.jpg",
                    "description": "The URL of the image to check. Must be publicly accessible."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": 1
                    },
                    "message": {
                      "type": "string",
                      "example": "succeeded"
                    },
                    "error": {
                      "type": "string",
                      "example": "none"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "watermark": {
                          "type": "object",
                          "properties": {
                            "detected": {
                              "type": "boolean",
                              "example": true,
                              "description": "Whether a visible watermark was detected"
                            },
                            "confidence": {
                              "type": "number",
                              "format": "float",
                              "example": 0.9,
                              "description": "Detection confidence between 0 and 1"
                            },
                            "source": {
                              "type": "string",
                              "nullable": true,
                              "example": "shutterstock",
                              "description": "Identified watermark source (stock agency, 'photographer', 'copyright', etc.) or null when undetermined"
                            },
                            "text": {
                              "type": "string",
                              "example": "shutterstock.com",
                              "description": "Short OCR snippet associated with the watermark (max 120 chars)"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "failed - unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unauthorized"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "failed - insufficient credits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "insufficient credits"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "failed - unknown server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "format": "int32",
                      "example": -1
                    },
                    "message": {
                      "type": "string",
                      "example": "failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "unknown server error"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiAuth": {
        "type": "apiKey",
        "description": "Your X-API-TOKEN is user id and api key seperated by a colon. Example: USERID:APIKEY",
        "name": "X-API-TOKEN",
        "in": "header"
      }
    }
  }
}